client/api/pyproject.toml
[project]
name = "dokanalyse-web-api"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.3",
    "async-lru>=2.3.0",
    "fastapi>=0.135.2",
    "lxml>=6.0.3",
    "python-levenshtein>=0.27.3",
    "python-multipart>=0.0.22",
    "shapely>=2.1.2",
    "types-lxml>=2026.2.16",
    "uvicorn>=0.42.0",
]
